+++ R/boost_tree.R
@@ -213,12 +213,13 @@ check_args.boost_tree <- function(object) {
   invisible(object)
 }
 
+
 # xgboost helpers --------------------------------------------------------------
 
 #' Boosted trees via xgboost
 #'
-#' `xgb_train` is a wrapper for `xgboost` tree-based models where all of the
-#'  model arguments are in the main function.
+#' `xgb_train()` and `xgb_predict()` are wrappers for `xgboost` tree-based
+#' models where all of the model arguments are in the main function.
 #'
 #' @param x A data frame or matrix of predictors
 #' @param y A vector (factor or numeric) or matrix (numeric) of outcome data.
@@ -251,16 +252,16 @@ check_args.boost_tree <- function(object) {
 #' @param event_level For binary classification, this is a single string of either
 #' `"first"` or `"second"` to pass along describing which level of the outcome
 #' should be considered the "event".
-#' @param ... Other options to pass to `xgb.train`.
+#' @param ... Other options to pass to `xgb.train()` or xgboost's method for `predict()`.
 #' @return A fitted `xgboost` object.
 #' @keywords internal
 #' @export
 xgb_train <- function(
-  x, y,
-  max_depth = 6, nrounds = 15, eta  = 0.3, colsample_bynode = NULL,
-  colsample_bytree = NULL, min_child_weight = 1, gamma = 0, subsample = 1,
-  validation = 0, early_stop = NULL, objective = NULL, counts = TRUE,
-  event_level = c("first", "second"), weights = NULL, ...) {
+    x, y, weights = NULL,
+    max_depth = 6, nrounds = 15, eta  = 0.3, colsample_bynode = NULL,
+    colsample_bytree = NULL, min_child_weight = 1, gamma = 0, subsample = 1,
+    validation = 0, early_stop = NULL, objective = NULL, counts = TRUE,
+    event_level = c("first", "second"), ...) {
 
   event_level <- rlang::arg_match(event_level, c("first", "second"))
   others <- list(...)
@@ -387,13 +388,17 @@ maybe_proportion <- function(x, nm) {
   }
 }
 
-xgb_pred <- function(object, newdata, ...) {
-  if (!inherits(newdata, "xgb.DMatrix")) {
-    newdata <- maybe_matrix(newdata)
-    newdata <- xgboost::xgb.DMatrix(data = newdata, missing = NA)
+#' @rdname xgb_train
+#' @param new_data A rectangular data object, such as a data frame.
+#' @keywords internal
+#' @export
+xgb_predict <- function(object, new_data, ...) {
+  if (!inherits(new_data, "xgb.DMatrix")) {
+    new_data <- maybe_matrix(new_data)
+    new_data <- xgboost::xgb.DMatrix(data = new_data, missing = NA)
   }
 
-  res <- predict(object, newdata, ...)
+  res <- predict(object, new_data, ...)
 
   x <- switch(
     object$params$objective,
@@ -470,6 +475,7 @@ get_event_level <- function(model_spec){
   event_level
 }
 
+
 #' @export
 #' @rdname multi_predict
 #' @param trees An integer vector for the number of trees in the ensemble.
@@ -500,9 +506,9 @@ multi_predict._xgb.Booster <-
   }
 
 xgb_by_tree <- function(tree, object, new_data, type, ...) {
-  pred <- xgb_pred(
+  pred <- xgb_predict(
     object$fit,
-    newdata = new_data,
+    new_data = new_data,
     iterationrange = c(1, tree + 1),
     ntreelimit = NULL
   )